+++ src/client/js/startup.js
@@ -3,27 +3,34 @@
  */
 
 Meteor.startup(() => {
-    if( CookieManager._conf.verbosity & CookieManager.C.Verbose.STORAGE ){
-        CookieManager.dump();
+    if( CookieManager._conf.verbosity & CookieManager.C.Verbose.DUMP ){
+        console.debug( 'CookieManager', CookieManager );
     }
+    console.debug( 'CookieManager', CookieManager );
+    if( CookieManager._conf.verbosity & CookieManager.C.Verbose.STORAGE ){
+        CookieManager.dumpStorage();
+    }
+    CookieManager.dumpStorage();
 });
 
 Meteor.startup(() => {
     // a technical cookie which stores the user consent as a composite (date,chosen action,list of enabled cookies)
     CookieManager.publish({
-        name: STORED_USER_CONSENT,
-        responsible: 'pwix:cookie-manager',
+        responsible: CM_RESPONSIBLE,
+        name: CM_USER_CONSENT_CK,
         description: pwixI18n.label( I18N, 'dialog.consent' ),
         category: CookieManager.C.Category.TECHNICALS,
         lifetime: CookieManager._conf.consentLifetime,
+        enabled: true,
         disableable: false
     });
     // fonctional cookies of the manager dialog width and height
     CookieManager.publish({
-        name: STORED_DIALOG_SIZE,
-        responsible: 'pwix:cookie-manager',
+        responsible: CM_RESPONSIBLE,
+        name: CM_DIALOG_SIZE_CK,
         description: pwixI18n.label( I18N, 'dialog.size' ),
         category: CookieManager.C.Category.FUNCTIONALS,
+        enabled: true,
         disableable: true
     });
 });