+++ src/common/js/global.js
@@ -3,6 +3,10 @@
  */
 
 CookieManager = {
+    // the published cookies
+    _published: [],
+    // the authorization status read from localStorage (null until reading)
+    _cookies: null
 };
 
 // make the CookieManager available to each and every package