+++ ChangeLog.md
@@ -8,7 +8,7 @@
 
     Release date: 
 
-    - 
+    - Remove cookies from our storage when user no more allow them (todo #25)
 
 ### 1.3.0